Address Matt's issues
* add a note about the new approach to the changelog
* no longer require tmux in the process of `make start`

+++ script/stop-server-in-tmux.sh
@@ -13,7 +13,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-if tmux has-session -t wellington; then
-    tmux kill-session -t wellington
+if hash tmux 2>/dev/null; then
+    if tmux has-session -t wellington; then
+        tmux kill-session -t wellington
+    fi
+else
+    echo "tmux not found; skipping tmux service step"
 fi
